Committing R2
Although line 10 can be combined with line 11, I thought I would split it just so that its a little easier to read.

We want to take the input from the user, and then find the difference between the second number and the first number. We then add that difference to the second number to find R2.