Committing Planina
This one was a little bit more complicated and took a little while longer, however when you break down the question it makes it a lot simplier.

If you can figure out the pattern in which the number of dots increase, you can then easily work out the number of dots at a given interation.

The pattern is this:

The number of dots is = to the starting number of dots square rooted, and increased iteratively by a rate of increase. The rate of increase starts at 1, and is doubled every iteration.