Committing Jumbo Javelin
We simply just want to take the first input and use that as our condition in our for loop

Then we first loop to get all the other javelin lengths, and then total all of those values.

But remember, the javelin length uses a 1cm after the 1st javelin, so we need to take that into account.