Committing Planetary Grid
Started to have a go at Planetary Grid, one of the hardest questions on Open Kattis.

My initial idea was first to introduce a Grid struct, which will represent our cartesian grid. Since we only care about all of the sections on the grid between us and the destination, we can set the bounds of our grid struct by the difference in X and Y values.

Then we can determine which cell in the grid is a road and not